+++ index.js
@@ -1,13 +1,12 @@
 // Node modules
 let fs = require('fs');
-let sass = require('sass');
 let express = require('express');
 let path = require('path');
-let DondaStatus = require('./public/models/donda-status');
+
+let sassService = require('./public/services/sass-service');
+let apiService = require('./public/services/spotify-api-service');
 
 // Fields
-const sass_input_path = 'public/stylesheets/main.scss';
-const css_output_path = 'public/stylesheets/main.css';
 const index_html_path = 'public/views/index.html';
 
 // Set up middleware
@@ -15,26 +14,7 @@ let app = express();
 app.use(express.static(path.join(__dirname, 'public')));
 
 // Render sass/scss
-sass.render(
-    {
-        file: sass_input_path,
-        outFile: css_output_path
-    },
-    (err, res) => {
-        if (err) {
-            console.error('Error rendering SCSS:');
-            console.error(err);
-            return;
-        } else {
-            fs.writeFile(css_output_path, res.css, (fs_err) => {
-                if (fs_err) {
-                    console.error('Error writing to CSS file:');
-                    console.error(fs_err);
-                }
-            });
-        }
-    }
-);
+sassService.renderSass();
 
 // Set up root GET
 app.get('/', (req, res) => {
@@ -51,10 +31,8 @@ app.get('/', (req, res) => {
         } else {
             res.write(data);
 
-            // Get DONDA status and (for now) display to console
-            let dondaStatus = new DondaStatus();
-            dondaStatus.updateReleased();
-            console.log(dondaStatus.released);
+            // Get DONDA status
+            apiService.checkDondaReleased();
         }
         res.end();
     });
